Added Plex music source
Added the ability to add plex servers as a music source. This required some fixes in the music_manager plugin, which then required fixes to be backported to local_music_source.

The player should be more stable as a whole as a race condition when changing the player state *should* have been fixed.