Added artist filtered queue additions
Added the ability to help filter what artist a song has to be from to help with libraries with songs with the same/similar names.

Also the Terminal plugin was fixed, and the fix pushed into the main APPLES repository. This bug caused problems if one word was enclosed by quotes.